feat(core-api): delete legacy serviceAccounts router from pops-api (Track M1 phase 5 PR 3)
The `core.serviceAccounts.{list,create,revoke}` admin router migrated
into pops-core-api in #2889 and became the canonical handler via the
nginx dispatcher in #2897. The legacy mount on pops-api was kept as a
fall-through until consumers were verified clean.

Audit: no shell/web caller references `trpc.core.serviceAccounts.*`.
The procedures are admin-only (CLI/MCP), never invoked from
`httpBatchLink`, so the `^/trpc/core\.serviceAccounts\.` rule routes
every real-world URL to pops-core-api — the legacy mount is dead code.

Removes:
- `apps/pops-api/src/modules/core/service-accounts/` (router, types,
  service tRPC tests).
- `serviceAccounts` from `coreRouter` in `modules/core/index.ts`.

Preserves the integration-level `protectedProcedure` scope-enforcement
coverage that was colocated with the deleted router by moving it into
`apps/pops-api/src/trpc-scope.test.ts`, next to the `trpc.ts` module
those tests actually exercise. The pure `hasScopeFor` logic continues
to live in `packages/core-db/src/__tests__/service-accounts.test.ts`.

Updates the nginx dispatcher comment to record that PR 3 landed — the
M1 PR 2 rule is now the only handler for `/trpc/core.serviceAccounts.*`.

Refs #2832, #2889, #2897.
