+++ apps/pops-shell/nginx.conf
@@ -134,7 +134,8 @@ server {
     #
     # `core.serviceAccounts.*` was migrated into pops-core-api (Track M1
     # PR 1, #2889). Admin-only (CLI/MCP), no batched callers — prefix
-    # match safe.
+    # match safe. Track M1 PR 3 deleted the legacy router from pops-api,
+    # so this rule is now the only handler for `/trpc/core.serviceAccounts.*`.
     location ~ ^/trpc/core\.serviceAccounts\. {
         set $core_dispatcher_upstream http://core-api:3001;
         proxy_pass $core_dispatcher_upstream;
